first pass of creating the ai chat page
This was tough to break up into separate commits,
so what happens here is a big abstraction of the PostCard
used on the homepage so that
we can use it on the chat page as well.

I've also created multiple Search components, but unfortunately there is
some prop drilling to get
all the posts data to represent the sourcs.